:recycle:Refact : 코드 리팩토링 #21  - 각 기능에 대한 모듈화
